LPD-84842 Inline the PortletResourcesUtil as it is only used by ResourceUtil
